requirements.txt
absl-py==0.12.0
aiohttp==3.7.4.post0
argparse==1.4.0
async-timeout==3.0.1
cachetools==4.2.1
cython==0.29.22
fsspec==0.8.7
future==0.18.2
grpcio==1.36.1
imbalanced-learn==0.8.0
markdown==3.3.4
multidict==5.1.0
numpy
oauthlib==3.1.0
pandas
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-slugify==4.0.1
pyyaml==5.3.1
requests-oauthlib==1.3.0
rsa==4.7.2
seaborn
text-unidecode==1.3
werkzeug==1.0.1
wget==3.2
yarl==1.6.3
transformers
nltk
pyenchant